services/todo/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic-settings==2.4.0
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
PyJWT==2.9.0
# テスト（コンテナでpytestを動かす場合）
pytest==8.3.3
httpx==0.27.2